Refactor data gen script to properly format date for cassandra, create mockData file to separate large chunks of static mock data from generation script
